add header with tailwind
