+++ client/src/auth/constants.js
@@ -1,3 +1,3 @@
-export const API_URL = "http://192.168.1.88:3001/api";
-// export const API_URL = "http://localhost:3001/api";
+// export const API_URL = "http://192.168.1.88:3001/api";
+export const API_URL = "http://localhost:3001/api";
 // export const API_URL = "https://fit-partner-server.onrender.com/api";
