+++ client/src/layout/MenuLayout.jsx
@@ -1,56 +1,25 @@
-import { Link } from "react-router-dom";
 import { useAuth } from "../auth/AuthProvider";
-import { API_URL } from "../auth/constants";
+import { ButtonLink } from "../components/ButtonLink";
+import { Header } from "./Header";
+import { Footer } from "./Footer";
+import {
+    FaLinkedin,
+    FaTelegramPlane,
+    FaGithubSquare
+} from 'react-icons/fa';
+import {FaComputer} from 'react-icons/fa6'
 
 
 export const MenuLayout = ({ children }) => {
 
-    const auth = useAuth();
-    const user = auth.getUser();
-    const email = user.email;
-
-    const handleLogout = async (e) => {
-        e.preventDefault();
-        try {
-            const response = await fetch(`${API_URL}/auth/logout`, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${auth.getRefreshToken()}`,
-                },
-            });
-            if(response.ok) {
-                auth.logout();
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    };
-
     return (
-        <>
-            <header>
-                <nav>
-                    <ul>
-                        <li>
-                            <Link to='/home'>Inicio</Link>
-                        </li>
-                        <li>
-                            <Link to='/profile'>Perfil</Link>
-                        </li>
-                        <li>
-                            <Link to='/profile'>{user ? email : ""}</Link>
-                        </li>
-                        <li>
-                            <a href="#" onClick={handleLogout}>Cerrar Sesión</a>
-                        </li>
-                    </ul>
-                </nav>
-            </header>
-
+        <div className="container max-w-full h-screen bg-gray-200">
+            <Header></Header>
             <main className="home">
                 {children}
             </main>
-        </>
+           <Footer></Footer>
+        </div >
+
     )
 };
